Refactors param executor
